fix: 검색 안되는 이슈 수정 #63
